Maze solver
This project is based on Maze Solver. We have created and successfully developed a program that implements Maze Traversal.
A maze can be viewed as a graph if we consider each intersection in the maze to be a vertex, and we add edges to the graph between adjacent intersection that is not blocked by a wall.  Our mazes will always have a starting point and an ending point. There is always exactly one path from one intersection to any other juncture in the maze.
Our mazes will have random weights specified between any two adjacent junctures. These values can be thought of as the "path" of traveling from one intersection to an adjacent one. The weights will be used for running Dijkstra's algorithm. The project will solve the mazes by first converting the maze into a graph, and then applying Dijkstra's algorithms:
Dijkstra's Algorithm: To find the path from "start" to "end" that uses edges with the shortest path. This only has ONE path from start to finish.